Complete signature ID tests migration
