+++ tests/integration/test_analysisd/test_signature_id/__init__.py
@@ -7,5 +7,5 @@ from pathlib import Path
 # Constants & base paths
 DATA_PATH = Path(Path(__file__).parent, 'data')
 RULES_SAMPLE_PATH = Path(DATA_PATH, 'rules_samples')
-CONFIGS_PATH = Path(DATA_PATH, 'configuration_templates')
+CONFIGS_PATH = Path(DATA_PATH, 'config_templates')
 TEST_CASES_PATH = Path(DATA_PATH, 'test_cases')
