+++ tests/integration/test_analysisd/test_signature_id/test_valid_signature_id.py
@@ -44,10 +44,10 @@ import pytest
 
 from pathlib import Path
 
-from wazuh_testing.constants.paths.logs import WAZUH_LOG_PATH
+from wazuh_testing.constants.paths.logs import OSSEC_LOG_PATH
 from wazuh_testing.modules.analysisd.testrule import patterns
-from wazuh_testing.tools import file_monitor
-from wazuh_testing.utils import callbacks, configuration
+from wazuh_testing.utils.config import load_configuration_template, get_test_cases_data
+from wazuh_testing.tools.file_monitor import FileMonitor, generate_callback
 
 from . import CONFIGS_PATH, TEST_CASES_PATH, RULES_SAMPLE_PATH
 
@@ -55,21 +55,18 @@ from . import CONFIGS_PATH, TEST_CASES_PATH, RULES_SAMPLE_PATH
 pytestmark = [pytest.mark.server, pytest.mark.tier(level=1)]
 
 # Configuration and cases data.
-configs_path = Path(CONFIGS_PATH, 'configuration_signature_id_values.yaml')
-test_cases_path = Path(TEST_CASES_PATH, 'cases_valid_signature_id.yaml')
+configs_path = Path(CONFIGS_PATH, 'config_signature_id_values.yaml')
+cases_path = Path(TEST_CASES_PATH, 'cases_valid_signature_id.yaml')
 
 # Test configurations.
-test_configuration, test_metadata, test_cases_ids = configuration.get_test_cases_data(test_cases_path)
-test_configuration = configuration.load_configuration_template(configs_path, test_configuration, test_metadata)
-
-# Test daemons to restart.
-daemons_handler_configuration = {'all_daemons': True}
+config_parameters, metadata, cases_ids = get_test_cases_data(cases_path)
+configuration = load_configuration_template(configs_path, config_parameters, metadata)
 
 
 # Test function.
-@pytest.mark.parametrize('test_configuration, test_metadata', zip(test_configuration, test_metadata), ids=test_cases_ids)
-def test_valid_signature_id(test_configuration, test_metadata, set_wazuh_configuration, truncate_monitored_files,
-                            prepare_custom_rules_file, daemons_handler):
+@pytest.mark.parametrize('configuration, metadata', zip(configuration, metadata), ids=cases_ids)
+def test_valid_signature_id(configuration, metadata, set_wazuh_configuration, truncate_monitored_files,
+                            prepare_custom_rules_file, restart_wazuh_function):
     '''
     description: Check that when a rule has an valid signature ID value assigned to the if_sid option, the rule is
                  not ignored.
@@ -83,7 +80,7 @@ def test_valid_signature_id(test_configuration, test_metadata, set_wazuh_configu
             - Check no log for "if_sid not found" is detected
             - Check no log for "empty if_sid" is detected
             - Check no log for "invalid if_sid" is detected
-        - Teardown:
+        - Tierdown:
             - Delete custom rule file
             - Restore configuration
             - Stop wazuh
@@ -93,10 +90,10 @@ def test_valid_signature_id(test_configuration, test_metadata, set_wazuh_configu
     tier: 1
 
     parameters:
-        - test_configuration:
+        - configuration:
             type: dict
-            brief: Configuration loaded from `configuration_templates`.
-        - test_metadata:
+            brief: Configuration loaded from `config_templates`.
+        - metadata:
             type: dict
             brief: Test case metadata.
         - set_wazuh_configuration:
@@ -108,9 +105,9 @@ def test_valid_signature_id(test_configuration, test_metadata, set_wazuh_configu
         - prepare_custom_rules_file:
             type: fixture
             brief: Copies custom rules_file before test, deletes after test.
-        - daemons_handler:
+        - restart_wazuh_function:
             type: fixture
-            brief: Handler of Wazuh daemons.
+            brief: Restart wazuh at the start of the module to apply configuration.
 
     assertions:
         - Check that wazuh starts
@@ -118,19 +115,14 @@ def test_valid_signature_id(test_configuration, test_metadata, set_wazuh_configu
         - Check ".*wazuh-testrule.*Empty 'if_sid' value. Rule '(\\d*)' will be ignored.*"
 
     input_description:
-        - The `configuration_signature_id_values.yaml` file provides the module configuration for
+        - The `config_signature_id_values.yaml` file provides the module configuration for
           this test.
         - The `cases_valid_signature_id.yaml` file provides the test cases.
     '''
     # Start monitors
-    monitor_not_found = file_monitor.FileMonitor(WAZUH_LOG_PATH)
-    monitor_not_found.start(callback=callbacks.generate_callback(patterns.SID_NOT_FOUND))
-
-    monitor_empty = file_monitor.FileMonitor(WAZUH_LOG_PATH)
-    monitor_empty.start(callback=callbacks.generate_callback(patterns.EMPTY_IF_SID_RULE_IGNORED))
-
-    monitor_invalid = file_monitor.FileMonitor(WAZUH_LOG_PATH)
-    monitor_invalid.start(callback=callbacks.generate_callback(patterns.INVALID_IF_SID_RULE_IGNORED))
+    monitor_not_found = FileMonitor(OSSEC_LOG_PATH, generate_callback(patterns.SID_NOT_FOUND))
+    monitor_empty = FileMonitor(OSSEC_LOG_PATH, generate_callback(patterns.EMPTY_IF_SID_RULE_IGNORED))
+    monitor_invalid = FileMonitor(OSSEC_LOG_PATH, generate_callback(patterns.INVALID_IF_SID_RULE_IGNORED))
 
     # Check that no log appears for rules if_sid field pointing to a non existent SID
     assert not monitor_not_found.callback_result
