--- tests/legacy/integration/pytest.ini
@@ -1,10 +1,0 @@
-[pytest]
-addopts = --strict-markers
-markers =
-    tier(level)
-    darwin
-    linux
-    sunos5
-    win32
-    server
-    agent
