--- tests/legacy/integration/test_analysisd/test_signature_id/data/rules_samples/valid_signature_id_rules.xml
@@ -1,10 +1,0 @@
-<group name="syscollector,">
-
-    <rule id="969696" level="3" >
-        <description>Initial rule</description>
-    </rule>
-    <rule id="969697" level="3" >
-        <if_sid>969696</if_sid>
-        <description>Custom Rule with valid if_sid</description>
-    </rule>
-</group>
