--- tests/legacy/integration/test_analysisd/test_signature_id/data/test_cases/cases_invalid_signature_id.yaml
@@ -1,13 +1,0 @@
-- name: Empty signature ID - Rule ignored
-  description: Check that if_sid is empty, a rule will be ignored
-  configuration_parameters:
-  metadata:
-    rules_file: empty_signature_id_rules.xml
-    is_empty: true
-
-- name: Invalid signature ID - Rule ignored
-  description: Check that if_sid is invalid, a rule will be ignored
-  configuration_parameters:
-  metadata:
-    rules_file: invalid_signature_id_rules.xml
-    is_empty: false
